+++ src/includes/hooks.inc.php
@@ -20,7 +20,9 @@ if(!defined('WPINC')) // MUST have WordPress.
 /*
 Add the plugin Actions/Filters here.
 */
-add_action('plugins_loaded', 'c_ws_plugin__s2member_translations::load');
+
+//260506 Translations loaded at init to avoid PHP Notice since WP 6.7.
+add_action('init', 'c_ws_plugin__s2member_translations::load');
 
 add_action('set_current_user', 'c_ws_plugin__s2member_user_securities::set_current_user', 2);